kit/kit_scripts/label_mask_to_coco/requirements.txt
certifi==2021.5.30
cycler==0.10.0
imageio==2.9.0
kiwisolver==1.3.2
matplotlib==3.4.3
networkx==2.6.3
numpy==1.21.2
opencv-python==4.5.3.56
Pillow==8.3.2
pyparsing==2.4.7
python-dateutil==2.8.2
PyWavelets==1.1.1
scikit-image==0.18.3
scipy==1.7.1
Shapely==1.7.1
six==1.16.0
tifffile==2021.8.30
